+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 [1.9.1]: https://github.com/GemsTracker/gemstracker-library/compare/1.9.0...1.9.1
 ### Added user functionality
 - Add a new track field type for linking tracks to other tracks (#621)
+- Add an external survey name to a survey (#625)
 - Added Url site management (#618)
 - Allow relation to ask for a token mail resend (#601)
 - Enable inserting more than one survey at a time (#617)